% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargepackages.R
\name{chargepackage}
\alias{chargepackage}
\title{Chargement des packages en mettant a jour R}
\usage{
chargepackage(df)
}
\arguments{
\item{df}{la liste des packages a installer. idealement a recuperer avec df <- installed.packages()}
}
\description{
Chargement des packages en mettant a jour R
}
\examples{
df <- installed.packages()
chargepackage(df)
}
