% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublons.R
\name{doublons}
\alias{doublons}
\title{doublons}
\usage{
doublons(df, ...)
}
\arguments{
\item{df}{la table}

\item{...}{lister les variables sur lesquelles on cherche les doublons}
}
\value{
la table des doublons avec l'ensemble des observations en doublon
}
\description{
A l'ensemble de la table d'origine, elle crée une variable Nbre_doublons qui
précise le nombre de doulons existants pour les variables choisies
}
\examples{
df <- flights
a <- doublons(df, origin, dest, time_hour)
a

}
