% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportcsv.R
\name{exportcsv}
\alias{exportcsv}
\title{exportcsv}
\usage{
exportcsv(p, append = FALSE, quote = TRUE, sep = ";", eol = "\\n",
  na = "NA", dec = ",", row.names = FALSE, col.names = TRUE,
  qmethod = c("escape", "double"), fileEncoding = "UTF-8")
}
\arguments{
\item{p}{le fichier à exporter}

\item{append}{booleen}

\item{quote}{booleen}

\item{sep}{separateur}

\item{eol}{je sais pas}

\item{na}{les valeurs de na}

\item{dec}{decimal des nombres}

\item{row.names}{booleen}

\item{col.names}{booleen}

\item{qmethod}{je sais pas}

\item{fileEncoding}{encoding}
}
\value{
un fichier en csv
}
\description{
Pour exporter un fichier il fallait deux arguments obligatoires
Avec cette fonction, il n'en faut plus qu'une : le nom de l'objet.
celui ci devient le nom du fichier à la sortie assorti de la date d'export
}
\examples{
a <- iris
exportcsv(a)
}
\author{
Lewis Hounkpevi
}
