% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportxlsx.R
\name{exportxlsx}
\alias{exportxlsx}
\title{exportxlsx}
\usage{
exportxlsx(p, sheetName = "Feuil1", col.names = T, row.names = F)
}
\arguments{
\item{p}{le nom du ficher à exporter}

\item{sheetName}{nom de l'onglet}

\item{col.names}{maintenair le header}

\item{row.names}{maitenir le row name}
}
\value{
l'objet exporté
}
\description{
Pour exporter un fichier il fallait deux arguments obligatoires
Avec cette fonction, il n'en faut plus qu'une : le nom de l'objet.
celui ci devient le nom du fichier à la sortie assorti de la date d'export
}
\examples{
a <- iris
exportcsv(a)
}
\author{
Lewis Hounkpevi
}
