% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysave.R
\name{mysave}
\alias{mysave}
\title{mysave}
\usage{
mysave(..., name)
}
\arguments{
\item{...}{citer des objets R comme dans save}

\item{name}{nom du fichier Rdata}
}
\value{
sauvegarder le ficher
}
\description{
Cette fonction est créée pour améliorer légèrement save.
elle ajoutera la date d'export au nom du fichier
}
\examples{
a <- c("a", "b", "c")
b <- c(1, 2, 3)
mysave(a, b, name = "sauvegarde")
}
\author{
Lewis Hounkpevi
}
