% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysaveall.R
\name{mysaveall}
\alias{mysaveall}
\title{mysaveall}
\usage{
mysaveall(name = "sauvegarde_")
}
\arguments{
\item{name}{nom de mon fichier}
}
\value{
sauvegarder une image
}
\description{
Cette fonction est creee pour améliorer legerement save.image.
elle ajoutera la date d'export au nom du fichier
}
\examples{
a <- c("a", "b", "c")
b <- c(1, 2, 3)
mysaveall()
}
\author{
Lewis Hounkpevi
}
