% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{resave}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{...}{les objets à resauvegarder}

\item{list}{maintenir ce parametre comme tel}

\item{file}{l'ancienne sauvegarde}
}
\description{
resave
}
\author{
Stackoverflow
}
